FIX 98b6d2f: write errors simplified by bufio.Writer
bufio.Writer does a bit of what I was achieving with xWriter:
stops subsequent writes if there was write error already - but does it
only after flush!
The difference from xWriter is that code doing writes will proceed
(as the writes are buffered), so no early returns in transformLine loop -
but this is good enough. After all, the amount of input data processed
in each line is not a lot.

The code gets much simpler and also, the fact that we use such buffered io
is made explicite - writeChunks* use bufio.Writer, not just any Writer.

So, the proc code goes back to what it was before 98b6d2f,
plus the actual error checking from Flush,
plus the above-mentioned explicit bufio.Writer arg.
What shines and is preserved from that original commit are failingWriter tests
(which got fixed by the way, see below).

Actually I am happy I discovered this pattern to buffer/flush not only the data,
but also errors, via bufio.Writer. This solves a lot of cases in other projects
where I would add extra checks for Write, or Fprint et al.

What else goes with this commit:
- fixed failingWriter so that it stops write in the middle if limit reached
- full proc tests, not only writeChunks

And now I have 100% coverage for the toplevel package!